Migrate ReportIssueController to report_issue Blueprint
